+++ server/migrations/1786827635631_player-name-team-unique.sql
@@ -1,0 +1,7 @@
+-- Up Migration
+
+ALTER TABLE player ADD CONSTRAINT player_name_team_uk UNIQUE (name, team);
+
+-- Down Migration
+
+ALTER TABLE player DROP CONSTRAINT player_name_team_uk;