+++ shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./roles";
 export * from "./league";
 export * from "./valuation";
+export * from "./player";